# internal/apidocs/openapi.yaml
openapi: "3.1.0"
info:
  title: VTA Farm API
  version: "1.0.0"
  description: |
    REST API for managing VTA setup sessions with per-user namespace isolation.

    ## Authentication

    All accounts use **passkey-only** authentication. There are no passwords.
    Protected endpoints use **httpOnly cookies** set at login.
    Cookie `vtafarm_user` is required for User endpoints; `vtafarm_admin` for Admin endpoints.

    ### In Scalar / browser
    1. Call the passkey login begin endpoint for your account type
    2. Complete the passkey challenge
    3. The browser stores the cookie automatically — all subsequent requests carry it

    ### curl / non-browser clients
    Use the `token` field from the login response body as a Bearer token:
    ```
    Authorization: Bearer <token>
    ```
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://vtafarm-api.firstperson.dev
    description: Production
...
components:
  securitySchemes:
    CookieAuthUser:
      type: apiKey
      in: cookie
      name: vtafarm_user
    CookieAuthAdmin:
      type: apiKey
      in: cookie
      name: vtafarm_admin
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
paths:
...
  /api/v1/admin/enroll/{token}:
    get:
      summary: "(public) Validate admin enrollment token"
      description: Checks whether the enrollment token is valid (not expired, not already used).
      tags: [Auth — Admin]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  expires_at:
                    type: string
                    format: date-time
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Token expired or already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: "(public) Enroll admin via token"
      description: |
        Consumes the enrollment token and sets the `vtafarm_admin` cookie (auto-login),
        so the admin can immediately register their passkey without a separate login step.
        Tokens are single-use and valid for 24 hours.
      tags: [Auth — Admin]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enrolled and logged in — vtafarm_admin cookie is set
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  unique_id:
                    type: string
                  token:
                    type: string
                    description: JWT for non-browser clients
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Token expired or already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Auth — User ─────────────────────────────────────────────────────────────
...
  /api/v1/admins:
    get:
      summary: List admins
      description: Returns all admin accounts, oldest first.
      tags: [Admin]
      security:
        - CookieAuthAdmin: []
      responses:
        "200":
          description: List of admins
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    unique_id:
                      type: string
                      example: "a1b2c3d4"
                    created_at:
                      type: string
                      format: date-time
                    updated_at:
                      type: string
                      format: date-time
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create admin enrollment token
      description: |
        Generates a 24-hour single-use enrollment token. Share the token with the new admin
        so they can call `POST /admin/enroll/{token}` to create their account and register
        their passkey. The admin account is created at enrollment time, not here.
      tags: [Admin]
      security:
        - CookieAuthAdmin: []
      responses:
        "201":
          description: Enrollment token created
          content:
            application/json:
              schema:
                type: object
                properties:
                  enrollment_token:
                    type: string
                    example: "a3f8b2c1d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1"
                  enrollment_expires:
                    type: string
                    format: date-time
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/invitations:
    post:
      summary: Create invitation link
      description: Generates a single-use invitation link valid for 24 hours.
      tags: [Admin — Invitations]
      security:
        - CookieAuthAdmin: []
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  token:
                    type: string
                    example: "a3f8b2c1d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1"
                  expires_at:
                    type: string
                    format: date-time
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: List invitation links
      description: Returns all invitation links, newest first.
      tags: [Admin — Invitations]
      security:
        - CookieAuthAdmin: []
      responses:
        "200":
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    token:
                      type: string
                    admin_id:
                      type: integer
                    expires_at:
                      type: string
                      format: date-time
                    used_at:
                      type: string
                      format: date-time
                      nullable: true
                    created_at:
                      type: string
                      format: date-time
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/invitations/{token}:
    get:
      summary: "(public) Validate invitation token"
      description: Checks whether the invitation token is valid (not expired, not already used).
      tags: [Admin — Invitations]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  expires_at:
                    type: string
                    format: date-time
        "404":
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Invitation expired or already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/setup/{id}/admin:
    post:
      summary: Provision admin DID
      description: |
        Phase 2 of setup: imports an admin DID into the VTA, then creates the VTA Deployment and Service.
        Session must be in `vta_setup_complete` status before calling this.
      tags: [User]
      security:
        - CookieAuthUser: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                admin_did:
                  type: string
                  description: DID to import as admin into the VTA
                  example: "did:key:z6Mk..."
              required: [admin_did]
      responses:
        "202":
          description: Provisioning started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: provisioning
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Session not in vta_setup_complete status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: K8s not configured on server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"